+++ apps/src/componentLibrary/button/_baseButton/_BaseButton.story.tsx
@@ -140,8 +140,8 @@ GroupOfColorsOf_BaseButtons.args = {
       onClick: () => null,
     },
     {
-      text: 'Button Secondary Grey',
-      color: buttonColors.grey,
+      text: 'Button Secondary Gray',
+      color: buttonColors.gray,
       type: 'secondary',
       size: 'm',
       onClick: () => null,
@@ -208,7 +208,7 @@ GroupOfColorsOf_BaseButtons.args = {
     },
     {
       icon: {iconName: 'smile', iconStyle: 'solid'},
-      color: buttonColors.grey,
+      color: buttonColors.gray,
       type: 'secondary',
       isIconOnly: true,
       size: 'm',